LogosShop#17 to dev

+++ src/main/java/com/vansisto/logosshop/service/impl/ImageServiceImpl.java
@@ -2,6 +2,8 @@ package com.vansisto.logosshop.service.impl;
 
 import com.vansisto.logosshop.domain.ImageDTO;
 import com.vansisto.logosshop.entity.Image;
+import com.vansisto.logosshop.exception.AlreadyExistsException;
+import com.vansisto.logosshop.exception.NotFoundException;
 import com.vansisto.logosshop.repository.ImageRepository;
 import com.vansisto.logosshop.service.ImageService;
 import com.vansisto.logosshop.util.ModelMapperUtil;
@@ -9,6 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class ImageServiceImpl implements ImageService {
@@ -21,38 +24,41 @@ public class ImageServiceImpl implements ImageService {
 
     @Override
     public ImageDTO create(ImageDTO dto) {
-        Image savedEntity = repository.save(map(dto));
-        return map(savedEntity);
+        if (!Objects.isNull(dto.getId()) && repository.existsById(dto.getId()))
+            throw new AlreadyExistsException("Image entity", "id", dto.getId());
+        return map(repository.save(map(dto)));
     }
 
     @Override
     public ImageDTO update(ImageDTO dto) {
-        Image updatedEnity = repository.save(map(dto));
-        return map(updatedEnity);
+        if (Objects.isNull(dto.getId()) && !repository.existsById(dto.getId()))
+            throw new NotFoundException("Image entity", "id", dto.getId());
+        return map(repository.save(map(dto)));
     }
 
     @Override
     public ImageDTO delete(ImageDTO dto) {
-        repository.delete(mapper.map(dto, Image.class));
+        repository.delete(map(dto));
         return dto;
     }
 
     @Override
     public String deleteById(String id) {
+        if (!repository.existsById(id)) throw new NotFoundException("Image entity", "id", id);
         repository.deleteById(id);
         return id;
     }
 
     @Override
     public ImageDTO getEntity(String id) {
-        return mapper.map(repository.getById(id), ImageDTO.class);
+        Image image = repository.findById(id).orElseThrow(() -> new NotFoundException("Image entity", "id", id));
+        return map(image);
     }
 
     @Override
     public List<ImageDTO> getAll() {
         return mapper.mapAll(repository.findAll(), ImageDTO.class);
     }
-
 
     private ImageDTO map(Image entity) {
         return mapper.map(entity, ImageDTO.class);
